Move ReplaceBody operation to setter category
The ReplaceBody operation was moved to the setter category and remaned
to HttpSetBody. This is a more intuitive name and location to look for.
